moving tlm packetizer configuration file
